+++ common/src/main/java/io/github/mortuusars/horseman/Config.java
@@ -1,6 +1,6 @@
 package io.github.mortuusars.horseman;
 
-import io.github.mortuusars.horseman.world.calling.CallDimensionHandling;
+import io.github.mortuusars.horseman.world.summoning.SummonDimensionHandling;
 import net.minecraft.resources.ResourceLocation;
 import net.neoforged.neoforge.common.ModConfigSpec;
 
@@ -31,14 +31,14 @@ public class Config {
         public static final ModConfigSpec.BooleanValue HORSE_HITCH_REQUIRES_LEAD;
         public static final ModConfigSpec.BooleanValue HORSE_HITCH_INVENTORY_SLOT;
 
-        // Calling
+        // Summoning
         public static final ModConfigSpec.IntValue COPPER_HORN_SOUND_RANGE;
         public static final ModConfigSpec.IntValue COPPER_HORN_COOLDOWN;
-        public static final ModConfigSpec.IntValue HORSE_CALLING_MAX_DISTANCE;
-        public static final ModConfigSpec.DoubleValue HORSE_CALLING_MAX_WALKING_DISTANCE;
-        public static final ModConfigSpec.DoubleValue HORSE_CALLING_WALK_MOVEMENT_SPEED;
-        public static final ModConfigSpec.EnumValue<CallDimensionHandling> HORSE_CALLING_DIMENSION_HANDLING;
-        public static final ModConfigSpec.ConfigValue<List<? extends String>> HORSE_CALLING_DIMENSIONS;
+        public static final ModConfigSpec.IntValue HORSE_SUMMONING_MAX_DISTANCE;
+        public static final ModConfigSpec.DoubleValue HORSE_SUMMONING_MAX_WALKING_DISTANCE;
+        public static final ModConfigSpec.DoubleValue HORSE_SUMMONING_WALK_MOVEMENT_SPEED;
+        public static final ModConfigSpec.EnumValue<SummonDimensionHandling> HORSE_SUMMONING_DIMENSION_HANDLING;
+        public static final ModConfigSpec.ConfigValue<List<? extends String>> HORSE_SUMMONING_DIMENSIONS;
 
         // Free Camera
         public static final ModConfigSpec.BooleanValue HORSE_FREE_CAMERA;
@@ -125,7 +125,7 @@ public class Config {
             }
 
             {
-                builder.push("calling");
+                builder.push("summoning");
 
                 COPPER_HORN_SOUND_RANGE = builder
                         .comment("Range in blocks where Copper Horn tooting sound can be heard by other players. Goat Horn has range of 256.")
@@ -136,31 +136,31 @@ public class Config {
                                 "If set to '-1' - will use same cooldown as Goat Horn.")
                         .defineInRange("copper_horn_cooldown", -1, -1, Integer.MAX_VALUE);
 
-                HORSE_CALLING_MAX_DISTANCE = builder
-                        .comment("Furthest distance in blocks from a player where a horse can still be called.",
+                HORSE_SUMMONING_MAX_DISTANCE = builder
+                        .comment("Furthest distance in blocks from a player where a horse can still be summoned.",
                                 "Set to -1 to allow any distance.")
-                        .defineInRange("max_calling_distance", -1, -1, Integer.MAX_VALUE);
+                        .defineInRange("max_summoning_distance", -1, -1, Integer.MAX_VALUE);
 
-                HORSE_CALLING_MAX_WALKING_DISTANCE = builder
+                HORSE_SUMMONING_MAX_WALKING_DISTANCE = builder
                         .comment("Furthest distance in blocks from a player from where a horse will walk instead of teleporting.",
                                 "Set to '0' to always teleport.",
                                 "Default: 32")
                         .defineInRange("max_walking_distance", 32.0, 0.0, 1024.0);
 
-                HORSE_CALLING_WALK_MOVEMENT_SPEED = builder
-                        .comment("Horse movement speed when it walks to the player when called.")
+                HORSE_SUMMONING_WALK_MOVEMENT_SPEED = builder
+                        .comment("Horse movement speed when it walks to the player when summoned.")
                         .defineInRange("walk_movement_speed", 2, 0.1, 2.5);
 
-                HORSE_CALLING_DIMENSION_HANDLING = builder
-                        .comment("Calling behavior depending on dimension.",
-                                "ANY: can call in any dimension.",
-                                "SAME: can only call when horse is in the same dimension as the player.",
-                                "WHITELIST: can only call when the player is in one of the dimensions defined in 'dimensions' list.",
-                                "BLACKLIST: can only call when the player is NOT in one of the dimensions defined in 'dimensions' list.",
+                HORSE_SUMMONING_DIMENSION_HANDLING = builder
+                        .comment("Summoning behavior depending on dimension.",
+                                "ANY: can summon in any dimension.",
+                                "SAME: can only summon when horse is in the same dimension as the player.",
+                                "WHITELIST: can only summon when the player is in one of the dimensions defined in 'dimensions' list.",
+                                "BLACKLIST: can only summon when the player is NOT in one of the dimensions defined in 'dimensions' list.",
                                 "Default: ANY")
-                        .defineEnum("dimension_handling", CallDimensionHandling.ANY);
+                        .defineEnum("dimension_handling", SummonDimensionHandling.ANY);
 
-                HORSE_CALLING_DIMENSIONS = builder
+                HORSE_SUMMONING_DIMENSIONS = builder
                         .comment("Dimensions whitelist or blacklist, depending on the 'dimension_handling' setting.",
                                 "Format: [\"minecraft:overworld\", \"minecraft:the_end\"]",
                                 "Default: [] (empty)")