+++ common/src/main/java/io/github/mortuusars/horseman/HorsemanServer.java
@@ -1,7 +1,7 @@
 package io.github.mortuusars.horseman;
 
 import com.google.common.base.Preconditions;
-import io.github.mortuusars.horseman.world.calling.HorseCalling;
+import io.github.mortuusars.horseman.world.summoning.Summoning;
 import net.minecraft.server.MinecraftServer;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -9,14 +9,14 @@ import org.jetbrains.annotations.Nullable;
 public class HorsemanServer {
     private static @Nullable MinecraftServer server;
 
-    private static @Nullable HorseCalling horseCalling;
+    private static @Nullable Summoning horseCalling;
 
     public static @NotNull MinecraftServer getServer() {
         Preconditions.checkNotNull(server, "Tried to retrieve server before it has initialized.");
         return server;
     }
 
-    public static @NotNull HorseCalling horseCalling() {
+    public static @NotNull Summoning summoning() {
         Preconditions.checkNotNull(horseCalling, "Tried to retrieve horseCalling before the server has initialized.");
         return horseCalling;
     }
@@ -25,7 +25,7 @@ public class HorsemanServer {
 
     public static void init(MinecraftServer server) {
         HorsemanServer.server = server;
-        horseCalling = new HorseCalling(server);
+        horseCalling = new Summoning(server);
     }
 
     public static void stop(MinecraftServer server) {