+++ common/src/main/java/io/github/mortuusars/horseman/world/calling/CallDimensionHandling.java
@@ -1,8 +1,0 @@
-package io.github.mortuusars.horseman.world.calling;
-
-public enum CallDimensionHandling {
-    ANY,
-    SAME,
-    WHITELIST,
-    BLACKLIST;
-}