+++ common/src/main/java/io/github/mortuusars/horseman/world/summoning/SummonDimensionHandling.java
@@ -1,0 +1,8 @@
+package io.github.mortuusars.horseman.world.summoning;
+
+public enum SummonDimensionHandling {
+    ANY,
+    SAME,
+    WHITELIST,
+    BLACKLIST;
+}